+++ app/controllers/organization_account_requests_controller.rb
@@ -1,6 +1,5 @@
 class OrganizationAccountRequestsController < ApplicationController
   skip_before_action :authenticate_user!
-  skip_before_action :verify_and_set_current_person
   skip_verify_authorized
 
   def new
