+++ app/controllers/registrations_controller.rb
@@ -17,7 +17,7 @@ class RegistrationsController < Devise::RegistrationsController
       if resource.persisted?
         resource.add_role(:adopter, Current.organization)
         # Create a new person unless it already exists
-        unless (Person.where(email: resource.email).any?)
+        unless (Person.exists?(email: resource.email))
           Person.create!(user_id: resource.id, first_name: resource.first_name, last_name: resource.last_name, email: resource.email)
         end
       end
